refactor(project-list): barrel imports for better readability.
