+++ client/src/features/project-list/StudentProjects.tsx
@@ -1,7 +1,4 @@
-import Header from './components/Header';
-import SearchBar from './components/SearchBar';
-import UpButton from './components/UpButton';
-import { ProjectList } from './components/ProjectList';
+import { Header, SearchBar, ProjectList, UpButton } from "./components";
 
 //#region Student Projects
 export default function StudentProjects() {
