+++ client/src/features/project-list/components/index.ts
@@ -1,0 +1,15 @@
+import Header from "./Header";
+import Paginator from "./Paginator";
+import ProjectCard from "./ProjectCards";
+import ProjectList from "./ProjectList";
+import SearchBar from "./SearchBar";
+import UpButton from "./UpButton";
+
+export {
+    Header,
+    Paginator,
+    ProjectCard,
+    ProjectList,
+    SearchBar,
+    UpButton
+};
